+++ pages/src/components/client/BreadcrumbsField.tsx
@@ -2,16 +2,19 @@
 
 import type { ArrayFieldClientComponent } from 'payload'
 
-import { ArrayField, Button, Drawer, useModal } from '@payloadcms/ui'
+import { ArrayField, Button, Drawer, useModal, useTranslation } from '@payloadcms/ui'
 
 import { BreadcrumbsIcon } from '../../icons/BreadcrumbsIcon.js'
-import { usePluginTranslation } from '../../utils/usePluginTranslations.js'
+import type {
+  PluginPagesTranslations,
+  PluginPagesTranslationKeys,
+} from '../../translations/index.js'
 
 const breadcrumbsModalSlug = 'breadcrumbs-drawer'
 
 export const BreadcrumbsFieldModalButton: React.FC = () => {
   const { toggleModal } = useModal()
-  const { t } = usePluginTranslation()
+  const { t } = useTranslation<PluginPagesTranslations, PluginPagesTranslationKeys>()
 
   return (
     <Button
@@ -19,17 +22,20 @@ export const BreadcrumbsFieldModalButton: React.FC = () => {
       icon={<BreadcrumbsIcon />}
       onClick={() => toggleModal(breadcrumbsModalSlug)}
       size="small"
-      tooltip={t('showBreadcrumbs')}
+      tooltip={t('@jhb.software/payload-pages-plugin:showBreadcrumbs')}
     />
   )
 }
 export const BreadcrumbsField: ArrayFieldClientComponent = (props) => {
   const { field, path } = props
-  const { t } = usePluginTranslation()
+  const { t } = useTranslation<PluginPagesTranslations, PluginPagesTranslationKeys>()
 
   return (
     <div className="field-type breadcrumbs-field-component">
-      <Drawer slug={breadcrumbsModalSlug} title={t('breadcrumbs')}>
+      <Drawer
+        slug={breadcrumbsModalSlug}
+        title={t('@jhb.software/payload-pages-plugin:breadcrumbs')}
+      >
         <div style={{ padding: '20px' }}>
           <ArrayField {...props} field={field} path={path} readOnly={true} />
         </div>