+++ pages/src/components/client/SlugFieldClient.tsx
@@ -1,4 +1,5 @@
 'use client'
+
 import type { TextFieldClientProps } from 'payload'
 
 import {
@@ -11,15 +12,19 @@ import {
   useField,
   useFormModified,
   useLocale,
+  useTranslation,
 } from '@payloadcms/ui'
 import { useCallback, useEffect, useState } from 'react'
 
 import { useCreateRedirect } from '../../hooks/useCreateRedirect.js'
 import { formatSlug, liveFormatSlug } from '../../hooks/validateSlug.js'
 import { RefreshIcon } from '../../icons/RefreshIcon.js'
+import type {
+  PluginPagesTranslations,
+  PluginPagesTranslationKeys,
+} from '../../translations/index.js'
 import { pathFromBreadcrumbs } from '../../utils/pathFromBreadcrumbs.js'
-import { usePluginTranslation } from '../../utils/usePluginTranslations.js'
-import { useBreadcrumbs } from '../client/hooks/useBreadcrumbs.js'
+import { useBreadcrumbs } from './hooks/useBreadcrumbs.js'
 
 export type SlugFieldProps = {
   defaultValue: Record<string, string> | string | undefined
@@ -40,7 +45,7 @@ export const SlugFieldClient = (clientProps: SlugFieldProps & TextFieldClientPro
   const [showSyncButtonTooltip, setShowSyncButtonTooltip] = useState(false)
   const { value: isRootPage } = useField<boolean>({ path: 'isRootPage' })
   const locale = useLocale()
-  const { t } = usePluginTranslation()
+  const { t } = useTranslation<PluginPagesTranslations, PluginPagesTranslationKeys>()
   const { createRedirect, isCreating, isSuccess } = useCreateRedirect(redirectsCollectionSlug)
   const { getBreadcrumbs } = useBreadcrumbs()
   const modified = useFormModified()
@@ -155,7 +160,7 @@ export const SlugFieldClient = (clientProps: SlugFieldProps & TextFieldClientPro
             >
               <>
                 <Tooltip show={showSyncButtonTooltip}>
-                  {t('syncSlugWithX').replace(
+                  {t('@jhb.software/payload-pages-plugin:syncSlugWithX').replace(
                     '{X}',
                     fallbackField.charAt(0).toUpperCase() + fallbackField.slice(1),
                   )}
@@ -199,7 +204,7 @@ export const SlugFieldClient = (clientProps: SlugFieldProps & TextFieldClientPro
               >
                 <div
                   dangerouslySetInnerHTML={{
-                    __html: t('slugWasChangedFromXToY')
+                    __html: t('@jhb.software/payload-pages-plugin:slugWasChangedFromXToY')
                       .replace('{X}', initialSlug)
                       .replace('{Y}', slug),
                   }}
@@ -215,7 +220,14 @@ export const SlugFieldClient = (clientProps: SlugFieldProps & TextFieldClientPro
                     onClick={handleCreateRedirect}
                     size="small"
                   >
-                    {t(isSuccess ? 'redirectCreated' : isCreating ? 'creating' : 'createRedirect')}
+                    {t(
+                      '@jhb.software/payload-pages-plugin:' +
+                        (isSuccess
+                          ? 'redirectCreated'
+                          : isCreating
+                            ? 'creating'
+                            : 'createRedirect'),
+                    )}
                   </Button>
                   <Button
                     buttonStyle="secondary"
@@ -224,7 +236,7 @@ export const SlugFieldClient = (clientProps: SlugFieldProps & TextFieldClientPro
                     onClick={() => setSlug(initialSlug)}
                     size="small"
                   >
-                    {t('revertSlug')}
+                    {t('@jhb.software/payload-pages-plugin:revertSlug')}
                   </Button>
                 </div>
               </div>